Create Points on a plane
Each point located on the plane can be described as a pair of coordinates customarily called x and y. We are going to write a Python class which stores both coordinates as float numbers. Moreover, we want the objects of this class to evaluate the distances between any of the two points situated on the plane.